test changes and cleanup
